% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHMdemo.R
\docType{data}
\name{CHMdemo}
\alias{CHMdemo}
\title{Canopy height model demo}
\format{A RasterLayer
\describe{Cell values are equal to canopy height above ground (in meters)}}
\usage{
CHMdemo
}
\description{
A small section of a canopy height model of a forest in British Columbia, Canada.
}
\keyword{datasets}

