% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm.R
\name{.calcGLCM}
\alias{.calcGLCM}
\title{Calculate GLCM}
\usage{
.calcGLCM(data, n_grey, angle = 0, d = 1, normalize = TRUE)
}
\description{
Some notes about this internal function:
1. Input should be a matrix
2. Shouldn't receive negative values
3. Shouldn't receive all NA values
4. Shouldn't be an empty matrix (i.e.: nrow = 0, ncol = 0)
5. 'n_grey' shouldn't be larger than the number of unique values
}
\details{
Remaining questions: What to do if there's only one value?
}
