% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestTools-package.R
\docType{package}
\name{ForestTools-package}
\alias{ForestTools}
\alias{ForestTools-package}
\title{ForestTools: Tools for Analyzing Remote Sensing Forest Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for analyzing remote sensing forest data, including functions for detecting treetops from canopy models, outlining tree crowns, and calculating textural metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrew-plowright/ForestTools}
  \item Report bugs at \url{https://github.com/andrew-plowright/ForestTools/issues}
}

}
\author{
\strong{Maintainer}: Andrew Plowright \email{andrew.plowright@alumni.ubc.ca}

Other contributors:
\itemize{
  \item Jean-Romain Roussel \email{jean-romain.roussel.1@ulaval.ca} (Contributed to segment-based GLCM segmentation) [contributor]
}

}
\keyword{internal}
