% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestTools.R, R/zzz.R
\docType{package}
\name{ForestTools}
\alias{ForestTools}
\title{Forest Tools: A package for analyzing remote sensing forest data}
\description{
Forest Tools provides functions for analyzing remote sensing forest data.
Functions like \link{vwf} and \link{mcws} are applied to rasterized
canopy height models (CHM) to detect treetops and outline their respective crowns.
}
\details{
Once a CHM is segmented into discrete crowns, \link{glcm} can compute textural metrics
for each crowns. This is usually done using orthorectified imagery, but can also
characterize the CHM's texture.
}
