% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glcm0}
\alias{glcm0}
\title{Create a 0 degree GLCM}
\usage{
glcm0(x, n_grey, d)
}
\arguments{
\item{x}{A Numeric matrix, integer values only}

\item{n_grey}{Number of grey levels}

\item{d}{distance from reference pixel to neighbour pixel}
}
\description{
Used internally by glcm()
}
