% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm.R
\name{glcm_img}
\alias{glcm_img}
\title{Get GLCM statistics for a single unsegmented image}
\usage{
glcm_img(img, n_grey = 32, angle = 0, d = 1, normalize = TRUE)
}
\arguments{
\item{img}{matrix or raster. Input image}

\item{n_grey}{integer. Number of grey levels used to discretize image}

\item{angle}{integer. Angle at which GLCM will be calculated. Valid inputs are 0, 45, 90, or 135}

\item{d}{numeric. Distance for calculating GLCM}

\item{normalize}{boolean. Normalize output image before calculating statistics}
}
\description{
Get GLCM statistics for a single unsegmented image
}
