% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kootenay.R
\docType{data}
\name{kootenayCrowns}
\alias{kootenayCrowns}
\title{Kootenay forest - Tree crowns}
\format{A \link[sp:SpatialPolygons]{SpatialPolygonsDataFrame} with the following attributes:
\describe{
  \item{height}{height of the tree's apex, in meters above ground. Inherited from \link{kootenayTrees}.}
  \item{winRadius}{radius of the moving window at the treetop's location. Inherited from \link{kootenayTrees}.}
  \item{crownArea}{area of crown outline in square meters}
  }}
\usage{
kootenayCrowns
}
\description{
Outlines of tree crowns corresponding to the \link{kootenayTrees} treetops. Generated using \link{mcws}.
}
\seealso{
\link{kootenayTrees} \link{kootenayCHM} \link{kootenayBlocks}
}
\keyword{datasets}
