% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kootenay.R
\docType{data}
\name{kootenayOrtho}
\alias{kootenayOrtho}
\title{Kootenay forest - Orthomosaic}
\format{A RasterLayer
\describe{Cell values are equal to canopy height above ground (in meters)}}
\source{
Data acquired from a photogrammetric drone survey performed by Spire Aerobotics
on June 16th, 2016.
}
\usage{
kootenayOrtho
}
\description{
An orthomosaic of a 1.5 hectare section of forest in the Kootenay mountains, in
British Columbia, Canada.
}
\seealso{
\link{kootenayTrees} \link{kootenayBlocks} \link{kootenayCrowns} \link{kootenayCHM}
}
\keyword{datasets}
