% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kootenay.R
\docType{data}
\name{kootenayTrees}
\alias{kootenayTrees}
\title{Kootenay forest - Dominant trees over 2 m}
\format{A \link[sp]{SpatialPointsDataFrame} with the following attributes:
\describe{
  \item{height}{height of the tree's apex, in meters above ground}
  \item{radius}{radius of the moving window (see \code{\link{TreeTopFinder}}) at
  the treetop's location}
  }}
\usage{
kootenayTrees
}
\description{
Dominant trees from a 1.5 hectare section of forest in the Kootenay mountains, in
British Columbia, Canada. Trees were detected by applying the \code{\link{TreeTopFinder}}
function to the \link{kootenayCHM} raster dataset. Only trees over 2 m above ground
were detected.
}
\seealso{
\link{kootenayCHM} \link{kootenayBlocks}
}
\keyword{datasets}

