% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quesnel.R
\docType{data}
\name{quesnelBlocks}
\alias{quesnelBlocks}
\title{Quesnel forest - Cut blocks}
\format{A \link[sp:SpatialPolygons]{SpatialPolygonsDataFrame} with the following attributes:
\describe{
  \item{BlockID}{numerical identifier for each block}
  \item{Shape_Leng}{length of polygon on meters}
  \item{Shape_Area}{area of polygon in square meters}
  }}
\usage{
quesnelBlocks
}
\description{
Boundaries of cut blocks within a 125 hectare section of forest in the Quesnel Timber Supply Area,
in British Columbia, Canada. Each block contains trees of different
levels of maturity. Overlaps with \link{quesnelTrees} and \link{quesnelCHM}.
}
\seealso{
\link{quesnelTrees} \link{quesnelCHM}
}
\keyword{datasets}
