% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quesnel.R
\docType{data}
\name{quesnelCHM}
\alias{quesnelCHM}
\title{Quesnel forest - Canopy height model}
\format{
A RasterLayer
\describe{Cell values are equal to canopy height above ground (in meters)}
}
\source{
Data acquired from a photogrammetric drone survey performed by Spire Aerobotics
on September 15th, 2016.
}
\usage{
quesnelCHM
}
\description{
A canopy height model of a 125 hectare section of forest in the Quesnel Timber Supply Area, in
British Columbia, Canada.
}
\seealso{
\link{quesnelTrees} \link{quesnelBlocks}
}
\keyword{datasets}
