% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quesnel.R
\docType{data}
\name{quesnelTrees}
\alias{quesnelTrees}
\title{Quesnel forest - Dominant trees over 2 m}
\format{A \link[sp]{SpatialPointsDataFrame} with the following attributes:
\describe{
  \item{height}{height of the tree's apex, in meters above ground}
  \item{winRadius}{radius of the moving window (see \code{\link{TreeTopFinder}}) at
  the treetop's location}
  }}
\usage{
quesnelTrees
}
\description{
Dominant trees from a 125 hectare section of forest in the Quesnel Timber Supply Area, in
British Columbia, Canada. Trees were detected by applying the \code{\link{TreeTopFinder}}
function to the \link{quesnelCHM} raster dataset. Only trees over 2 m above ground
were detected.
}
\seealso{
\link{quesnelCHM} \link{quesnelBlocks}
}
\keyword{datasets}

