% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestTools.R
\docType{package}
\name{ForestTools}
\alias{ForestTools}
\title{Forest Tools: A package for analyzing geospatial forest data}
\description{
Forest Tools provides functions for analyzing remotely sensed forest data.
Functions like \link{vwf} and \link{mcws} are applied to rasterized
canopy height models, and can detect treetops and outline their respective crowns. \link{sp_summarise} can
summarize tree counts and attributes within particular areas of interest or within continuous spatial
grids.
}
