% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwcs.R
\name{mcws}
\alias{mcws}
\title{Marker-Controlled Watershed Segmentation}
\usage{
mcws(
  treetops,
  CHM,
  minHeight = 0,
  format = "raster",
  OSGeoPath = NULL,
  verbose = FALSE
)
}
\arguments{
\item{treetops}{The point locations of treetops in sf format. The function will generally produce a
number of crown segments equal to the number of treetops.}

\item{CHM}{Canopy height model in SpatRaster format. Should be the same that was used to create
the input for \code{treetops}.}

\item{minHeight}{numeric. The minimum height value for a \code{CHM} pixel to be considered as part of a crown segment.
All \code{CHM} pixels beneath this value will be masked out. Note that this value should be lower than the minimum
height of \code{treetops}.}

\item{format}{string. Format of the function's output. Can be set to either 'raster' or 'polygons'.}

\item{OSGeoPath}{character. Obsolete. Will be removed next version}

\item{verbose}{logical. Print processing progress to console.}
}
\value{
Depending on the argument set with \code{format}, this function will return a map of outlined
crowns as either a SpatRaster class object, in which distinct crowns are given a unique cell value, or a sf class object, in which each crown
is represented by a polygon.
}
\description{
This function implements the \link[imager]{watershed} function to segment (i.e.: outline) crowns from a CHM (canopy height model).
Segmentation is guided by the point locations of treetops, typically detected using the \link{vwf} function.
See Meyer & Beucher (1990) for details on watershed segmentation.
}
\details{
Crown segments are returned as either a SpatRaster or a sf (Simple Feature) class object,
as defined using the \code{format} argument. For many analytic purposes, it is preferable to have
crown outlines as polygons. However, polygonal crown maps take up significantly more disk space, and take
longer to process. It is advisable to run this function using a raster output first, in order to check
its results and adjust parameters.

NOTE: when setting \code{format} to 'polygons', orphaned segments (i.e.: outlines without an associated treetop) will be removed.
This is an issue with the 'raster' format that has yet to be resolved.
}
\examples{
\dontrun{
# Use variable window filter to detect treetops in demo canopy height model
ttops <- vwf(CHMdemo, winFun = function(x){x * 0.06 + 0.5}, minHeight = 2)

# Use 'mcws' to outline tree crowns
segs <- mcws(ttops, CHMdemo, minHeight = 1)
}

}
\references{
Meyer, F., & Beucher, S. (1990). Morphological segmentation. \emph{Journal of visual communication and
image representation, 1}(1), 21-46.
}
\seealso{
\code{\link{vwf}} \code{\link{sp_summarise}} \code{\link[imager]{watershed}} \cr \cr
OSGeo4W download page: \url{https://trac.osgeo.org/osgeo4w/}
}
